+++ app/src/main/java/com/example/blackbox/domain/repository/RecordedUsageStatsRepository.kt
@@ -5,7 +5,7 @@ import kotlinx.coroutines.flow.Flow
 
 interface RecordedUsageStatsRepository {
     fun getAllRecordedUsageStats(): Flow<List<RecordedUsageStats>>
-    suspend fun getRecordedUsageStatsById(id: Int): RecordedUsageStats?
+    suspend fun getRecordedUsageStatsById(id: Long): RecordedUsageStats?
     suspend fun insertRecordedUsageStats(recordedUsageStats: RecordedUsageStats): Long
     suspend fun updateRecordedUsageStats(recordedUsageStats: RecordedUsageStats)
     suspend fun deleteRecordedUsageStats(recordedUsageStats: RecordedUsageStats)
